+++ src/com/landawn/abacus/util/BooleanPair.java
@@ -58,8 +58,8 @@ public class BooleanPair {
         return predicate.test(this) ? Optional.of(this) : Optional.<BooleanPair> empty();
     }
 
-    public Stream<BooleanPair> stream() {
-        return Stream.of(this);
+    public Stream<Boolean> stream() {
+        return Stream.of(_1, _2);
     }
 
     @Override