+++ src/com/landawn/abacus/util/Sheet.java
@@ -21,7 +21,6 @@ import java.util.Arrays;
 import java.util.Collection;
 import java.util.Iterator;
 import java.util.LinkedHashMap;
-import java.util.LinkedHashSet;
 import java.util.List;
 import java.util.Map;
 import java.util.NoSuchElementException;
@@ -58,8 +57,8 @@ public final class Sheet<R, C, E> implements Cloneable {
     }
 
     public Sheet(Collection<R> rowKeySet, Collection<C> columnKeySet) {
-        this._rowKeySet = new LinkedHashSet<>(rowKeySet);
-        this._columnKeySet = new LinkedHashSet<>(columnKeySet);
+        this._rowKeySet = N.newLinkedHashSet(rowKeySet);
+        this._columnKeySet = N.newLinkedHashSet(columnKeySet);
     }
 
     public Sheet(Collection<R> rowKeySet, Collection<C> columnKeySet, Object[][] rows) {