+++ src/com/landawn/abacus/util/ShortTriple.java
@@ -1,6 +1,6 @@
 package com.landawn.abacus.util;
 
-import com.landawn.abacus.util.stream.Stream;
+import com.landawn.abacus.util.stream.ShortStream;
 
 public class ShortTriple {
     public final short _1;
@@ -83,8 +83,8 @@ public class ShortTriple {
         return predicate.test(this) ? Optional.of(this) : Optional.<ShortTriple> empty();
     }
 
-    public Stream<ShortTriple> stream() {
-        return Stream.of(this);
+    public ShortStream stream() {
+        return ShortStream.of(_1, _2, _3);
     }
 
     @Override